requirements.txt
pandas>=2.0
numpy>=1.24
PyYAML>=6.0
pydantic>=2.0
jinja2>=3.1
python-dateutil>=2.9
pytest>=8.0
